requirements.txt
# ─── 必须 ──────────────────────────────────────────────────────
requests>=2.28.0        # AI API 调用（anthropic / openai / gemini / qwen）
imapclient>=2.3.0       # IMAP IDLE 实时推送模式
markdown>=3.4.0         # 邮件 Markdown 转 HTML 渲染
ddgs>=0.1.0               # DuckDuckGo 爬虫搜索（无需 API Key）
googlesearch-python>=1.2.3  # Google 爬虫搜索（无需 API Key）
croniter>=1.3.0         # cron 表达式解析（高级定时任务）
pystray>=0.19.5         # 系统托盘图标
Pillow>=10.0.0          # pystray 图标绘制依赖

# ─── Gmail OAuth（可选，使用 gmail + oauth_google 时需要）───────
# google-auth>=2.0.0
# google-auth-oauthlib>=1.0.0
# google-auth-httplib2>=0.1.0

# ─── Outlook OAuth（可选，使用 outlook + oauth_microsoft 时需要）
# msal>=1.20.0

# ─── Web UI（可选，运行 bash manage.sh webui 时自动安装）────────
# fastapi>=0.110.0
# uvicorn[standard]>=0.29.0
# jinja2>=3.1.0
# python-multipart>=0.0.9
# itsdangerous>=2.1.0
